FaceTrackAI/backend/requirements.txt
Flask==2.2.3
Flask-Cors==3.0.10
Werkzeug==2.2.3
numpy==1.24.2
Pillow==9.4.0
openai==1.12.0
soundfile==0.12.1
supabase==1.2.0
